[iter-1/step-10] lam: quality gate — ruff/mypy clean, full test suite green
2522/2522 pytest passing (11 skipped, 0 failed). Fixes applied:
- Math-notation noqa (RUF002/RUF003) on planner/executor/oracle + test_planner
- Explicit int/float casts in planner arithmetic for mypy
- model_name coercion in LiteLLMOracle for str typing
- FieldExtractionRequest: embed schema hint in prompt (drop invalid kwarg)
- Rename shadowing `l` test locals to `lf`; escape regex pattern in test_oracle

+++ src/fsm_llm/lam/planner.py
@@ -1,3 +1,4 @@
+# ruff: noqa: RUF002
 from __future__ import annotations
 
 """
@@ -154,7 +155,7 @@ def _predicted_calls(k: int, d: int) -> int:
     """
     if d == 0:
         return 1
-    return k**d
+    return int(k**d)
 
 
 def _predicted_cost(n: int, tau: int, k: int, d: int, alpha: float, c: float) -> float:
@@ -162,7 +163,7 @@ def _predicted_cost(n: int, tau: int, k: int, d: int, alpha: float, c: float) ->
     plus the leaf-level cost of ``k^d`` Leaf calls on inputs of size ≤ tau.
     """
     if d == 0:
-        return c * (n**alpha)
+        return float(c * (n**alpha))
     total = 0.0
     for level in range(d):
         num_nodes = k**level
@@ -171,7 +172,7 @@ def _predicted_cost(n: int, tau: int, k: int, d: int, alpha: float, c: float) ->
     # Leaf level: k^d pieces, each ≤ tau.
     leaf_size = n / (k**d)
     total += (k**d) * c * (leaf_size**alpha)
-    return total
+    return float(total)
 
 
 def _accuracy_floor(d: int, leaf_acc: float, combine_acc: float) -> float:
